update_backend_1h57-8/1

+++ controllers/songController.js
@@ -37,22 +37,26 @@ export const post_addNewSong = async (req, res) => {
     const {song_name, artist_name, album_name, song_image, song_info, duration, category} = req.body
     
     try {
-        var artist = await pool.query('SELECT artist.artist_id FROM artist,song WHERE artist_name = $1 and artist.artist_id = song.artist_id  LIMIT 1', [artist_name])
-        if(artist) {
-            if(album_name == null) {
-                var song = await pool.query('INSERT INTO song(song_id, artist_id, album_id, song_name, song_image, song_info, duration, category, average_rate, last_updated_stamp, created_stamp) \
-                VALUES(default, (SELECT artist_id FROM artist WHERE artist_name = $1 LIMIT 1), null, $3, $4, $5, $6, $7, null, null, default) RETURNING *', [artist_name, album_name, song_name, song_image, song_info, duration, category])
+        var name = await pool.query('SELECT song_name FROM song WHERE song_name = $1', [song_name])
+        if(name.rows[0].song_name == null) {
+            var artist = await pool.query('SELECT artist.artist_id FROM artist,song WHERE artist_name = $1 and artist.artist_id = song.artist_id  LIMIT 1', [artist_name])
+            if(artist) {
+                if(album_name == null) {
+                    var song = await pool.query('INSERT INTO song(song_id, artist_id, album_id, song_name, song_image, song_info, duration, category, average_rate, last_updated_stamp, created_stamp) \
+                    VALUES(default, (SELECT artist_id FROM artist WHERE artist_name = $1 LIMIT 1), null, $3, $4, $5, $6, $7, null, null, default) RETURNING *', [artist_name, album_name, song_name, song_image, song_info, duration, category])
+                }
+                else {
+                    var album = await pool.query('SELECT album.album_id FROM album, song WHERE album_name = $1 and album.album_id = song.album_id  LIMIT 1', [album_name])
+                    if(album) {
+                        var song = await pool.query('INSERT INTO song(song_id, artist_id, album_id, song_name, duration, category, average_rate, last_updated_stamp, created_stamp) \
+                            VALUES(default, (SELECT artist_id FROM artist WHERE artist_name = $1  LIMIT 1), (SELECT album_id FROM album WHERE album_name = $2 LIMIT 1), $3, $4, $5, $6, $7, null, null, default) RETURNING *', [artist_name, album_name, song_name, song_image, song_info, duration, category])
+                    }        
+                    else res.status(500).send({message: 'Album is not exist'});
+                }
             }
-            else {
-                var album = await pool.query('SELECT album.album_id FROM album, song WHERE album_name = $1 and album.album_id = song.album_id  LIMIT 1', [album_name])
-                if(album) {
-                    var song = await pool.query('INSERT INTO song(song_id, artist_id, album_id, song_name, duration, category, average_rate, last_updated_stamp, created_stamp) \
-                        VALUES(default, (SELECT artist_id FROM artist WHERE artist_name = $1  LIMIT 1), (SELECT album_id FROM album WHERE album_name = $2 LIMIT 1), $3, $4, $5, $6, $7, null, null, default) RETURNING *', [artist_name, album_name, song_name, song_image, song_info, duration, category])
-                }        
-                else res.status(500).send({message: 'Album is not exist'});
-            }
+            else res.status(500).send({message: 'Artist unknown'});
         }
-        else res.status(500).send({message: 'Artist unknown'});
+        else res.status(500).send({message: 'Song name already exists'})
     } catch (err) {
         console.log(err.stack)
     }
@@ -71,25 +75,29 @@ export const post_deleteSong = async (req, res) => {
     console.log(req.body)
     const {id} = req.body
     try {
-        var song = pool.query('DELETE FROM song WHERE song_id = $1', [id])
-        console.log('delete')
-        if (song) {
-            res.status(201).send({message: 'Song deleted', data: song.rows});
-            var artist = await pool.query('SELECT artist_name FROM artist, song WHERE artist.song_id = album.song_id')
-            if (artist) {
-                res.status(201).send({message: 'Update number of songs in artist successful'})
-                artistUpdateNumofSongs(artist.rows[0].artist_name)
+        var playlist = await pool.query('DELETE FROM song_added_to_playlist WHERE song_id = $1', [id])
+        if(playlist) {            
+            var song = await pool.query('DELETE FROM song WHERE song_id = $1', [id])
+            console.log('delete')
+            if (song) {
+                res.status(201).send({message: 'Song deleted', data: song.rows});
+                var artist = await pool.query('SELECT artist_name FROM artist, song WHERE artist.song_id = album.song_id')
+                if (artist) {
+                    res.status(201).send({message: 'Update number of songs in artist successful'})
+                    artistUpdateNumofSongs(artist.rows[0].artist_name)
+                }
+                else res.status(500).send({message: 'Error in updating number of songs in artist'})
+                var album = await pool.query('SELECT album_name FROM song, album WHERE album.song_id = album.song_id')
+                if (album) {
+                    res.status(201).send({message: 'Update number of songs in album successful'})
+                    if(album.rows[0] != null) albumUpdateNumofSongs(album.rows[0].album_name)
+                }
+                else res.status(500).send({message: 'Error in updating number of songs in album'})
+            } else {
+                res.status(500).send({message: 'Error in deleting song'});
             }
-            else res.status(500).send({message: 'Error in updating number of songs in artist'})
-            var album = await pool.query('SELECT album_name FROM song, album WHERE album.song_id = album.song_id')
-            if (album) {
-                res.status(201).send({message: 'Update number of songs in album successful'})
-                if(album.rows[0] != null) albumUpdateNumofSongs(album.rows[0].album_name)
-            }
-            else res.status(500).send({message: 'Error in updating number of songs in album'})
-        } else {
-            res.status(500).send({message: 'Error in deleting song'});
         }
+        else  res.status(500).send({message: 'Error in deleting song in playlist'});
     } catch (err) {
         console.log(err.stack)
     }
